refactor: improve Day class
Improve the `Day` class by updating it to work well with MyFitnessPal
(there appear to have been changes since the original code was written)
and to make it return in a more useable format.

Also increment the version number by a hint, add myself as an author,
and add Pry because I'm an addict and can't live without it.